FEND-954: rebuild bramch. remove fake commit
